Daemonify the script to enable a painless global install.
